Making add return and expect doubles

+++ Small/Small.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int add(int x, int y)
+double add(double x, double y)
 {
 	return x + y;
 }
